fix(migrate): address Holly r1 findings on #117 (#115 Phase 2)
S1 — drop dead `piDevSkillId` import + companion `void` line from
`claude-skills-substrate-verify.ts`; Pi.dev id collisions can't occur
on the per-skill verify surface, so the leftover-from-cut-path
reference adds noise without value.

S2 — extract `parseDescriptionFromFrontmatter` + `stripQuotes` +
`FRONTMATTER_RE` into new `src/claude-skills-frontmatter.ts`. Both
the migrator (source-side description extraction) and the verifier
(projected-side parser) now share one implementation, eliminating
the duplicated ~20-line scanner pair.

N1 — collapse a stray double-blank line after `migrateClaudeSkills`.

N2 — tighten the unmapped-path smoke test: rewrite the stream-of-
consciousness fixture comment to a one-line statement of the realistic
outcome, and replace the `expect([...]).toContain(status)` widening
with a strict `expect(status).toBe("verified-with-warnings")` now
that the contract is settled.
